back-end/HappyBack/bot/requirements.txt
amqp==1.4.9
anyjson==0.3.3
astroid==1.5.3
billiard==3.3.0.23
celery==3.1.25
certifi==2017.11.5
chardet==3.0.4
Django==1.11.7
django-celery==3.2.2
djangorestframework==3.7.3
idna==2.6
isort==4.2.15
kombu==3.0.37
lazy-object-proxy==1.3.1
mccabe==0.6.1
powerline-status==2.6
psycopg2==2.7.3.2
pylint==1.7.4
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
vine==1.1.4
virtualenv==15.1.0
wrapt==1.10.11
